backend/requirements.txt
pip==22.3.1
requests
psycopg2-binary==2.9.9
fastapi[standard]
starlette
pydantic